Boring

+++ include/GUI.hpp
@@ -18,7 +18,7 @@ public:
   GUIContainer *get_root_container() const;
   void add_container(GUIContainer *container);
   void add_text(GUIText *text);
-  void on_window_resize(double width, double height);
+  void on_window_resize(unsigned width, unsigned height);
   void on_mouse_move(double x1, double y1, double x2, double y2);
   void on_mouse_down(double x, double y);
   void on_mouse_up(double x, double y);
